Moved the database config files to the backend, the database file is now stored in the db folder.
