+++ apps/backend/src/db/drizzle/meta/_journal.json
@@ -5,8 +5,8 @@
     {
       "idx": 0,
       "version": "6",
-      "when": 1753742407602,
-      "tag": "0000_messy_thaddeus_ross",
+      "when": 1753871422087,
+      "tag": "0000_exotic_roulette",
       "breakpoints": true
     }
   ]
