+++ apps/backend/src/db/types/fastify.d.ts
