--- packages/db/tsconfig.json
@@ -1,24 +1,0 @@
-
-{
-	"compilerOptions": {
-		"target": "es2022",
-		"module": "commonjs",
-		"moduleResolution": "node",
-		"rootDir": "src",
-		"outDir": "./dist",
-		"esModuleInterop": true,
-		"forceConsistentCasingInFileNames": true,
-		"strict": true,
-    "skipLibCheck": true,
-    "types": ["node"]
-	},
-	"include": [
-    "src/**/*",
-    "types/**/*"
-  ],
-	"exclude": [
-    "node_modules",
-    "dist",
-    "drizzle.config.ts"
-  ]
-}
